refactor(rolls): extract roll-types module
refs #559
